dom/tests/mochitest/general/test_frameElementWrapping.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for location object behaviors</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
 
</div>
<iframe id="ifr" src="file_frameElementWrapping.html"></iframe>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

var count = 0;

function runTest(result, message) {
    ok(result === 'PASS', message);

    if (++count === 2)
        SimpleTest.finish();
    else
        $('ifr').contentWindow.location = 'http://example.org/tests/dom/tests/mochitest/general/file_frameElementWrapping.html';
}

window.addEventListener("message",
                        function(event) { runTest.apply(null, event.data.split(',')) },
                        false);

</script>
</pre>
</body>
</html>
